Separate Commands and Queries into separate folders
