+++ src/TournamentManagement/TournamentManagement.Application/Commands/AddTournamentCommand.cs
@@ -6,21 +6,19 @@ using TournamentManagement.Contract;
 using TournamentManagement.Domain.TournamentAggregate;
 using TournamentManagement.Domain.VenueAggregate;
 
-namespace TournamentManagement.Application
+namespace TournamentManagement.Application.Commands
 {
-	public class AmendTournamentCommand : ICommand
+	public sealed class AddTournamentCommand : ICommand
 	{
-		public TournamentId Id { get; }
 		public string Title { get; }
 		public TournamentLevel TournamentLevel { get; }
 		public DateTime StartDate { get; }
 		public DateTime EndDate { get; }
 		public VenueId VenueId { get; }
 
-		public AmendTournamentCommand(Guid tournamentGuid, string title, TournamentLevel level, DateTime startDate,
+		public AddTournamentCommand(string title, TournamentLevel level, DateTime startDate,
 			DateTime endDate, Guid venueGuid)
 		{
-			Id = new TournamentId(tournamentGuid);
 			Title = title;
 			TournamentLevel = level;
 			StartDate = startDate;
@@ -29,41 +27,36 @@ namespace TournamentManagement.Application
 		}
 	}
 
-	public sealed class AmendTournamentCommandHandler : ICommandHandler<AmendTournamentCommand>
+	public sealed class AddTournamentCommandHandler : ICommandHandler<AddTournamentCommand, Guid>
 	{
 		private readonly IUnitOfWork _uow;
 
-		public AmendTournamentCommandHandler(IUnitOfWork uow)
+		public AddTournamentCommandHandler(IUnitOfWork uow)
 		{
 			_uow = uow;
 		}
 
-		public Result Handle(AmendTournamentCommand command)
+		public Result<Guid> Handle(AddTournamentCommand command)
 		{
-			var tournament = _uow.TournamentRepository.GetById(command.Id);
-			if (tournament == null)
-			{
-				return Result.Failure("Tournament does not exist");
-			}
-
 			var venue = _uow.VenueRepository.GetById(command.VenueId);
 			if (venue == null)
 			{
-				return Result.Failure<TournamentId>("Venue does not exist");
+				return Result.Failure<Guid>("Venue does not exist");
 			}
 
 			try
 			{
-				tournament.UpdateDetails(command.Title, command.TournamentLevel, 
+				var tournament = Tournament.Create(command.Title, command.TournamentLevel,
 					command.StartDate, command.EndDate, venue);
 
+				_uow.TournamentRepository.Add(tournament);
 				_uow.SaveChanges();
 
-				return Result.Success(tournament.Id);
+				return Result.Success(tournament.Id.Id);
 			}
 			catch (Exception ex)
 			{
-				return Result.Failure<TournamentId>(ex.Message);
+				return Result.Failure<Guid>(ex.Message);
 			}
 		}
 	}
