+++ src/TournamentManagement/TournamentManagement.WebApi/Controllers/TournamentController.cs
@@ -3,6 +3,8 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using TournamentManagement.Application;
+using TournamentManagement.Application.Commands;
+using TournamentManagement.Application.Queries;
 using TournamentManagement.Contract;
 
 namespace TournamentManagement.WebApi.Controllers
